requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-socketio==5.5.1
eventlet==0.37.0
anthropic>=0.52.0
pillow>=11.1.0
pillow-heif>=0.21.0
python-dotenv==1.0.1
gunicorn>=22.0.0
